PlayerTest

+++ ClassLibrary1/libCPP/Algo.cpp
@@ -13,7 +13,7 @@ using namespace std;
 void Algo::fillMap(TileType map[], int size)
 {
 	// je considere size comme la taille du côté de la carte
-	int nbMaxType = pow(size, 2)/4;
+	double nbMaxType = pow(size, 2)/4;
 	vector<int> max;
 	for (int i = 0; i < 4; i++)
 	{
